% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanData.R
\name{cleanData}
\alias{cleanData}
\title{Clean helminth parasite occurrence data}
\usage{
cleanData(data, speciesOnly = FALSE, validateHosts = FALSE)
}
\arguments{
\item{data}{helminth interaction data}

\item{speciesOnly}{boolean flag to remove host and parasite species
where data are only available at genus level (default = FALSE)}

\item{validateHosts}{boolean flag to check host species names
against Catalogue of Life information and output taxonomic
information (default = FALSE)}
}
\value{
data.frame with cleaned data
}
\description{
Given the data or a subset of the helminth interaction data 
(see `loadData()`), perform various cleaning functions on the
 data, including validation of species names, run taxize on 
hosts (again, as there are already host and parasite taxonomic
variables as part of this data.frame , and remove 
records only to genus level.
}
\details{
Use \code{data(locations)} for a list of possible locations.
}
\author{
Tad Dallas
}
