% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listLocations.R
\name{listLocations}
\alias{listLocations}
\title{Lists geographic locations}
\usage{
listLocations()
}
\value{
Returns a vector of locations. However, it is important to note that
some locations are nested inside of other locations. For instance, you can
search for all host-parasite associations in the Great Lakes as a whole, or
you can target Lake Ontario specifically.
}
\description{
Lists geographic locations that can be input to `findHost()` or
`findParasite()` and the corresponding latitude and longitude coordinates 
of the country's centroid. The georeferencing is performed dynamically using 
`ggmap` and is potentially incorrect.
}
\examples{

\dontrun{listLocations()}


}
\references{
Gibson, D. I., Bray, R. A., & Harris, E. A. (Compilers) (2005).
Host-Parasite Database of the Natural History Museum, London.
}
\author{
Tad Dallas
}
