% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadData.R
\name{loadData}
\alias{loadData}
\title{Load the helminth interaction data}
\usage{
loadData(outdir = ".")
}
\arguments{
\item{outdir}{output directory (default is the top-level of the package repo)}
}
\value{
data.frame of host-helminth interactions
}
\description{
Load the helminth interaction data
}
\references{
Gibson, D. I., Bray, R. A., & Harris, E. A. (Compilers) (2005).
Host-Parasite Database of the Natural History Museum, London.
}
\author{
Tad Dallas
}
