% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locations.R
\docType{data}
\name{locations}
\alias{locations}
\title{Table of geographic location names, and associated coordinates}
\format{\describe{
   \item{Location}{Name of geographic location}
   \item{Latitude}{Latitude of location centroid}
   \item{Longitude}{Longitude of location centroid}
 }}
\usage{
data(locations)
}
\description{
A cached copy of the output of listLocations().
}
\references{
Gibson, D. I., Bray, R. A., & Harris, E. A. (Compilers) (2005).
Host-Parasite Database of the Natural History Museum, London.
}
\keyword{data}
