% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/findHost.R
\name{findHost}
\alias{findHost}
\title{Find parasite occurrence data for given host.}
\usage{
findHost(genus, species = "", location = "", citation = FALSE)
}
\arguments{
\item{genus}{Host genus}

\item{species}{Host species}

\item{location}{Geographic location.}

\item{citation}{Boolean. Should the output include the citation link? default is FALSE}
}
\value{
Three (or four) column data.frame containing host species, parasite species
(shortened name and full name), and citation link (optional), with each row
corresponding to an occurrence
of a parasite species on a host species.
}
\description{
Given a host genus, species, and/or location, returns a list of parasite
occurrences on that host or for that location. Search available locations
using the `listLocations` function.
}
\details{
`location` values can be listed using the `listLocations()` function.
}
\examples{
gorillaParasites=findHost('Gorilla', 'gorilla')
head(gorillaParasites)
}
\author{
Tad Dallas
}
\references{
Gibson, D. I., Bray, R. A., & Harris, E. A. (Compilers) (2005).
Host-Parasite Database of the Natural History Museum, London. URL.
}
\seealso{
\code{\link{findParasite}}
}

