% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/findLocation.R
\name{findLocation}
\alias{findLocation}
\title{Find host-parasite interactions for a given location}
\usage{
findLocation(location = "", citation = FALSE)
}
\arguments{
\item{location}{Location of host-parasite interaction.}

\item{citation}{Boolean. Should the output include the citation link? default is FALSE}
}
\value{
Three (or four) column data.frame containing host species, parasite species
(shortened name and full name), and citation link (optional), with each row
corresponding to an occurrence
of a parasite species on a host species.
}
\description{
Given a location (available from `listLocations()`) this function returns
all host-parasite associations in that location.
}
\details{
Use the `listLocations()` function for a list of possible locations.
}
\examples{
FrenchHostPars=findLocation(location='France')
dim(FrenchHostPars)
}
\author{
Tad Dallas
}
\references{
Gibson, D. I., Bray, R. A., & Harris, E. A. (Compilers) (2005).
Host-Parasite Database of the Natural History Museum, London. URL.
}
\seealso{
\code{\link{findHost}}
}

