% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/findParasite.R
\name{findParasite}
\alias{findParasite}
\title{Find host-parasite interactions for a given parasite species.}
\usage{
findParasite(genus, species = "", location = "", citation = FALSE)
}
\arguments{
\item{genus}{Parasite genus}

\item{species}{Parasite species}

\item{location}{Location of host-parasite interaction.}

\item{citation}{Boolean. Should the output include the citation link? default is FALSE}
}
\value{
Three (or four) column data.frame containing host species, parasite species
(shortened name and full name), and citation link (optional), with each row
corresponding to an occurrence
of a parasite species on a host species.
}
\description{
Given a host genus and/or species, this function returns a matrix containing
host-parasite interaction data.
}
\details{
Use the `listLocations()` function for a list of possible locations.
}
\examples{
strongHosts=findParasite(genus='Strongyloides')
dim(strongHosts)
}
\author{
Tad Dallas
}
\references{
Gibson, D. I., Bray, R. A., & Harris, E. A. (Compilers) (2005).
Host-Parasite Database of the Natural History Museum, London. URL.
}
\seealso{
\code{\link{findHost}}
}

