% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helminthR-package.R
\docType{package}
\name{helminthR-package}
\alias{helminthR-package}
\alias{helminthR}
\title{Access London Natural History Museum host-helminth record database}
\description{
'helminthR': A programmatic interface to the London Natural History Museum's Host-Parasite database.

The package currently allows you to query by host species, parasite species, 
and geographic location. No information is provided on parasite prevalence or intensity.
}
\author{
Tad Dallas \email{tdallas@ucdavis.edu}
}
